pyproject.toml
[tool.poetry]
name = "customer-support-chat"
version = "0.1.0"
description = "Customer Support Chat Application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
langgraph = "^0.2.19"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.23"
ipython = "^8.27.0"
duckduckgo-search = "^6.3.0"
langchain-community = "^0.2.16"
langchain-google-community = "^1.0.8"
qdrant-client = "^1.11.1"
pandas = "^2.2.2"
aiohttp = "^3.10.5"
tqdm = "^4.66.5"
more-itertools = "^10.5.0"
pywin32 = {version = "306", optional = true, markers = "sys_platform == 'win32'"}
sentence-transformers = "^5.1.2"
google-generativeai = "^0.8.5"
dateparser = "^1.2.2"
python-telegram-bot = "^22.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
